Split JAXB engine into a separate module
