+++ pippo-jaxb/src/main/java/ro/pippo/jaxb/JaxbInitializer.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2014 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ro.pippo.jaxb;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ro.pippo.core.Application;
+import ro.pippo.core.Initializer;
+
+/**
+ * @author James Moger
+ */
+public class JaxbInitializer implements Initializer {
+
+    private static final Logger log = LoggerFactory.getLogger(JaxbInitializer.class);
+
+    @Override
+    public void init(Application application) {
+        application.registerContentTypeEngine(JaxbEngine.class);
+    }
+
+    @Override
+    public void destroy(Application application) {
+    }
+
+}
