+++ pippo-jaxb/src/main/resources/META-INF/services/ro.pippo.core.ContentTypeEngine
@@ -1,0 +1,1 @@
+ro.pippo.jaxb.JaxbEngine
