+++ pippo-jaxb/src/main/resources/pippo.properties
@@ -1,0 +1,1 @@
+initializer=ro.pippo.jaxb.JaxbInitializer
