+++ pom.xml
@@ -163,10 +163,11 @@
         <module>pippo-trimou</module>
         <module>pippo-spring</module>
         <module>pippo-guice</module>
-		<module>pippo-weld</module>
+        <module>pippo-weld</module>
         <module>pippo-gson</module>
         <module>pippo-fastjson</module>
         <module>pippo-jackson</module>
+        <module>pippo-jaxb</module>
         <module>pippo-xstream</module>
         <module>pippo-snakeyaml</module>
         <module>pippo-metrics</module>
